requirements.txt
yfinance>=0.2.40
pandas>=2.0
jinja2>=3.1
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
PyYAML>=6.0
pytz>=2024.1
python-dateutil>=2.9
